pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-cov>=6.1.1",
    "pyyaml>=6.0.2",
    "flask>=2.2.2",
    "gunicorn>=20.1.0",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.6",
    "requests>=2.32.4",
    "beautifulsoup4>=4.13.4",
    "validators>=0.35.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.8",
]
